pyproject.toml
[project]
name = "sample-mcp-server-skills"
version = "0.1.0"
description = "MCP Server for managing user skills via filesystem"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.4.1",
    "python-dotenv>=1.0.1",
    "pydantic>=2.0.0",
    "httpx>=0.20.0",
    "google-api-python-client>=2.118.0",
]

[project.scripts]
sample-mcp-server-skills = "src.main:mcp"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
]
